+++ isaaclab_arena_gr00t/utils/wait_for_gr00t_server.py
@@ -13,7 +13,7 @@ import time
 
 import zmq
 
-from isaaclab_arena_gr00t.utils.groot_path import ensure_gr00t_importable
+from isaaclab_arena_gr00t.utils.gr00t_path import ensure_gr00t_importable
 
 ensure_gr00t_importable()
